requirements.txt
openai>=1.0.0
gitpython>=3.1.0
pyyaml>=6.0
requests>=2.28.0
pathlib>=1.0.0
python-dotenv>=1.0.0